admin/requirements.txt
Django>=3.2
djangorestframework>=3.12
mysqlclient>=2.0.1
django-mysql>=3.9
django-cors-headers>=3.5.0
pika==1.2.0